+++ src/main/java/com/mobble/mobbleserver/account/user/CustomUserDetails.java
@@ -6,16 +6,13 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
 import java.util.Collection;
-import java.util.stream.Collectors;
+import java.util.Collections;
 
 public record CustomUserDetails(Member member) implements UserDetails {
 
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
-        // Member 엔티티가 가진 ClubMember 목록에서 Role을 가져와 Spring Security의 권한(GrantedAuthority)으로 변환합니다.
-        return member.getClubMembers().stream()
-                .map(clubMember -> new SimpleGrantedAuthority("ROLE_" + clubMember.getRole().name()))
-                .collect(Collectors.toList());
+        return Collections.singletonList(new SimpleGrantedAuthority("ROLE_USER"));
     }
 
     @Override
